Moving AST tests to dedicated folders
